requirements.txt
Django==5.2.8
djangorestframework==3.16.1
dotenv==0.9.9
openai==2.8.0
